Change useCurrencyField to useHandleCurrencyField and create new hook useCurrencyField
